requirements.txt
# Direct dependencies — versions match a tested Windows venv (Python 3.11+).
# Install: pip install -r requirements.txt
fastapi==0.136.1
uvicorn[standard]==0.46.0
python-multipart==0.0.27
pydantic==2.13.3
pydantic-settings==2.14.0
python-dotenv==1.2.2
SQLAlchemy==2.0.49
psycopg[binary]==3.3.3
alembic==1.18.4
passlib[bcrypt]==1.7.4
# bcrypt 4.1+ removes __about__ — passlib 1.7.4 breaks; see passlib bcrypt backend.
bcrypt>=4.0.1,<4.1
PyJWT==2.12.1
cryptography==47.0.0
boto3==1.42.97
email-validator==2.3.0
cuid==0.4
